# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "ActivityPage": {
        "description": "Newest first. Reverse client-side to show oldest-to-newest.",
        "properties": {
          "items": {
            "items": {
              "discriminator": {
                "mapping": {
                  "comment": "#/components/schemas/CommentActivity",
                  "due_date_changed": "#/components/schemas/DueDateChangedActivity",
                  "evaluation_closed": "#/components/schemas/EvaluationClosedActivity",
                  "evaluation_reopened": "#/components/schemas/EvaluationReopenedActivity",
                  "evaluation_submitted": "#/components/schemas/EvaluationSubmittedActivity",
                  "evaluator_added": "#/components/schemas/EvaluatorAddedActivity",
                  "evaluator_removed": "#/components/schemas/EvaluatorRemovedActivity",
                  "idea_created": "#/components/schemas/IdeaCreatedActivity",
                  "idea_edited": "#/components/schemas/IdeaEditedActivity",
                  "owner_changed": "#/components/schemas/OwnerChangedActivity",
                  "status_changed": "#/components/schemas/StatusChangedActivity"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/CommentActivity"
                },
                {
                  "$ref": "#/components/schemas/IdeaCreatedActivity"
                },
                {
                  "$ref": "#/components/schemas/IdeaEditedActivity"
                },
                {
                  "$ref": "#/components/schemas/StatusChangedActivity"
                },
                {
                  "$ref": "#/components/schemas/OwnerChangedActivity"
                },
                {
                  "$ref": "#/components/schemas/EvaluatorAddedActivity"
                },
                {
                  "$ref": "#/components/schemas/EvaluatorRemovedActivity"
                },
                {
                  "$ref": "#/components/schemas/EvaluationSubmittedActivity"
                },
                {
                  "$ref": "#/components/schemas/EvaluationClosedActivity"
                },
                {
                  "$ref": "#/components/schemas/EvaluationReopenedActivity"
                },
                {
                  "$ref": "#/components/schemas/DueDateChangedActivity"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ActivityPage",
        "type": "object"
      },
...
      "CommentActivity": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who did it; null if the user no longer exists."
          },
          "comment": {
            "$ref": "#/components/schemas/CommentBody"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idea_id": {
            "format": "uuid",
            "title": "Idea Id",
            "type": "string"
          },
          "type": {
            "const": "comment",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "idea_id",
          "created_at",
          "actor",
          "type",
          "comment"
        ],
        "title": "CommentActivity",
        "type": "object"
      },
...
      "CommentCreate": {
        "additionalProperties": false,
        "properties": {
          "body_md": {
            "description": "Markdown.",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body Md",
            "type": "string"
          }
        },
        "required": [
          "body_md"
        ],
        "title": "CommentCreate",
        "type": "object"
      },
      "CommentUpdate": {
        "additionalProperties": false,
        "properties": {
          "body_md": {
            "description": "Markdown.",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body Md",
            "type": "string"
          }
        },
        "required": [
          "body_md"
        ],
        "title": "CommentUpdate",
        "type": "object"
      },
...
      "CurrentUser": {
        "description": "The signed-in user (``GET /auth/me``), also used for the dev login picker.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "description": "One or two letters for the avatar fallback.",
            "readOnly": true,
            "title": "Initials",
            "type": "string"
          },
          "is_platform_admin": {
            "title": "Is Platform Admin",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "display_name",
          "avatar_url",
          "email",
          "is_platform_admin",
          "initials"
        ],
        "title": "CurrentUser",
        "type": "object"
      },
      "DevLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "DevLoginRequest",
        "type": "object"
      },
...
      "FieldError": {
        "description": "One request-validation error.",
        "properties": {
          "loc": {
            "description": "Location, e.g. ['body', 'title'].",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details, returned as ``application/problem+json``.\n\nClients should branch on ``code`` (stable, snake_case), not on ``title`` or\n``detail`` (human-readable, may change). As RFC 9457 allows, ``detail``,\n``instance`` and ``request_id`` may be absent: the one exception to \"every\ndocumented response field is always present\".",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ValidationProblem": {
        "description": "422 problem with per-field errors (``code`` is ``validation_error``).",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ValidationProblem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Server-side session cookie set by sign-in (Phase 1: POST /api/v1/auth/dev/login). Unsafe methods must also send the X-CSRF-Token header equal to the soundings_csrf cookie.",
        "in": "cookie",
        "name": "soundings_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Soundings REST API.\n\n* Errors are [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) problem details\n  (`application/problem+json`); branch on the `code` field.\n* Lists use cursor pagination: pass `next_cursor` back as `?cursor=`.\n* JSON fields are snake_case; timestamps are ISO 8601 with a UTC offset.\n",
    "title": "Soundings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/dev/login": {
      "post": {
        "description": "Starts a session for the user: sets the soundings_session (HttpOnly) and soundings_csrf cookies. Development only: 404 unless SOUNDINGS_DEV_LOGIN_ENABLED=true (always refused in production).",
        "operationId": "dev_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "summary": "Sign in as a user without a password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/dev/users": {
      "get": {
        "description": "Active users, platform admins first, then by name. Development only: 404 unless SOUNDINGS_DEV_LOGIN_ENABLED=true (always refused in production).",
        "operationId": "list_dev_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CurrentUser"
                  },
                  "title": "Response List Dev Users",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "summary": "List users for the dev login picker",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Ends the session (if any) and clears the session cookies. Always 204.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "summary": "Sign out",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "The signed-in user. 401 when not signed in: show the sign-in page.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Who am I",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/comments/{comment_id}": {
      "delete": {
        "description": "The author (403 not_author) or a project admin. Leaves a 'comment deleted' placeholder in the feed.",
        "operationId": "delete_comment",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Comment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "activity"
        ]
      },
      "patch": {
        "description": "The author only (403 not_author); sets edited_at. Deleted comments are 404.",
        "operationId": "update_comment",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Comment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentActivity"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Edit a comment",
        "tags": [
          "activity"
        ]
      }
    },
...
    "/api/v1/ideas/{idea}/activity": {
      "get": {
        "description": "Comments and events, newest first. Never contains scores.",
        "operationId": "list_idea_activity",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          },
          {
            "description": "Opaque cursor from next_cursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from next_cursor.",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad request (e.g. invalid_cursor)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Activity feed",
        "tags": [
          "activity"
        ]
      }
    },
    "/api/v1/ideas/{idea}/comments": {
      "post": {
        "description": "Members and admins. Returns the new feed item; the author watches the idea.",
        "operationId": "create_comment",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentActivity"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Comment on an idea",
        "tags": [
          "activity"
        ]
      }
    },
...
  }
}